join introspection
