+++ src/Data/Morpheus/Resolve/Operator.hs
@@ -19,7 +19,7 @@ module Data.Morpheus.Resolve.Operator
   ) where
 
 import           Data.Morpheus.Resolve.Encode               (ObjectFieldResolvers (..), resolveBySelection, resolversBy)
-import           Data.Morpheus.Resolve.Generics.TypeRep     (ObjectRep (..), TypeUpdater, resolveTypes)
+import           Data.Morpheus.Resolve.Introspect           (ObjectRep (..), TypeUpdater, resolveTypes)
 import           Data.Morpheus.Schema.SchemaAPI             (hiddenRootFields, schemaAPI, schemaTypes)
 import           Data.Morpheus.Types.Internal.AST.Selection (SelectionSet)
 import           Data.Morpheus.Types.Internal.Data          (DataArguments, DataFingerprint (..), DataType (..),
